web_app/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
Django==3.1.5
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-summernote==0.8.11.6
fake-useragent==0.1.11
idna==2.10
lxml==4.6.2
parse==1.19.0
Pillow==8.1.0
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
python-slugify==4.0.1
pytz==2020.5
requests==2.25.1
requests-html==0.10.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
tqdm==4.56.0
urllib3==1.26.2
w3lib==1.22.0
websockets==8.1
psycopg2-binary
gunicorn
